+++ R/data.R
@@ -17,11 +17,12 @@
 #' @source \url{https://fred.stlouisfed.org/series/VIXCLS}
 "logVIX"
 
-#' University of Michigan: inflation expectation time series: IE
+#' Spread between 10-Year and 1-Year treasury rates: T10Y1Y
 #'
-#' A dataset containing monthly inflation expectations from 1078I to 2018XII.
+#' A dataset containing monthly U.S. interest rate spread between the 10-Year Treasury constant
+#' maturity rate and 1-Year Treasury constant maturity rate from 1953IV to 2020II.
 #'
-#' @format A class 'ts' time series object containing 492 observations.
+#' @format A class 'ts' time series object containing 803 observations.
 #'
-#' @source \url{https://fred.stlouisfed.org/series/MICH}
-"IE"
+#' @source \url{https://fred.stlouisfed.org/series/GS10} \url{https://fred.stlouisfed.org/series/GS1}
+"T10Y1Y"